requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
cryptography==44.0.3
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
executing==2.2.0
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
PyGithub==2.6.1
pygments==2.19.1
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.20.2
